refactor existing code into matrix package to support addition of points, regions packages later
