+++ starspace/matrix/write.py
